[BUGFIX] ReadSpeaker Widget not rendered if *_INT cObjects are used
If COA_INT or USER_INT cObjects are used somewhere in the page's
TypoScript configuration the ReadSpeaker Widget is not rendered
in TYPO3 6.2. The reason for that can be found in some changes
in the rendering pipeline leading to the fact that the content
being processed in the PageRenderer is not used anymore in case
those *_INT cObjects are used.

Resolves: #66841
Forge: https://forge.typo3.org/issues/66841